refactor: ModalContainer에서 포커스 트랩 기능을 useFocusTrap 훅으로 분리
