+++ components/src/lib/components/common/Modal.Container.tsx
@@ -1,8 +1,7 @@
-import { useEffect, useRef } from 'react';
-
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 
+import useFocusTrap from '../../hooks/useFocusTrap';
 import { useModalContext } from '../../hooks/useModalContext';
 
 type ModalContainerProps = {
@@ -34,41 +33,7 @@ const ModalContainer = ({
 }: ModalContainerProps) => {
   const { $zIndex = 1000, ...props } = useModalContext();
 
-  const modalRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key !== 'Tab' || !modalRef.current) {
-        return;
-      }
-
-      const focusableElements = modalRef.current.querySelectorAll<HTMLElement>(
-        'a[href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), button:not([disabled]), [tabindex]:not([tabindex="-1"])'
-      );
-      const firstElement = focusableElements[0];
-      const lastElement = focusableElements[focusableElements.length - 1];
-
-      if (!firstElement || !lastElement) {
-        return;
-      }
-
-      if (e.shiftKey) {
-        if (document.activeElement === firstElement) {
-          e.preventDefault();
-          lastElement.focus();
-        }
-      } else {
-        if (document.activeElement === lastElement) {
-          e.preventDefault();
-          firstElement.focus();
-        }
-      }
-    };
-    document.addEventListener('keydown', handleKeyDown);
-    return () => {
-      document.removeEventListener('keydown', handleKeyDown);
-    };
-  }, []);
+  const modalRef = useFocusTrap<HTMLDivElement>();
 
   return (
     <StyledModalContainer
