+++ components/src/lib/hooks/useFocusTrap.ts
@@ -1,0 +1,44 @@
+import { useEffect, useRef } from 'react';
+
+const useFocusTrap = <T extends HTMLElement>() => {
+  const ref = useRef<T>(null);
+
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key !== 'Tab' || !ref.current) {
+        return;
+      }
+
+      const focusableElements = ref.current.querySelectorAll<HTMLElement>(
+        'a[href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), button:not([disabled]), [tabindex]:not([tabindex="-1"])'
+      );
+      const firstElement = focusableElements[0];
+      const lastElement = focusableElements[focusableElements.length - 1];
+
+      if (!firstElement || !lastElement) {
+        return;
+      }
+
+      if (e.shiftKey) {
+        if (document.activeElement === firstElement) {
+          e.preventDefault();
+          lastElement.focus();
+        }
+      } else {
+        if (document.activeElement === lastElement) {
+          e.preventDefault();
+          firstElement.focus();
+        }
+      }
+    };
+
+    document.addEventListener('keydown', handleKeyDown);
+    return () => {
+      document.removeEventListener('keydown', handleKeyDown);
+    };
+  }, []);
+
+  return ref;
+};
+
+export default useFocusTrap;
